Don't flip the board based on "player name" when toggling analysis.
Automatic board flip based on "player name" is only supposed to happen when
loading a new game or similar. If you for some reason want to have the board
flipped compared to the "player name" orientation, it is annoying that the board
flips back every time you toggle analysis mode.