Use SVG graphics to draw chess pieces.
Use SVG graphics to draw chess pieces instead of a custom font. This should fix
the problem where some android devices ignore the requested font so the chess
pieces are drawn as two overlapping regular letters.

This should also make it easier to add alternative piece sets in the future.