Revert "Use SwitchPreference"
Change back to using CheckBoxPreference instead of SwitchPreference in
DroidFish. In Android 4.x the switch can be so big that the text does
not fit, since it uses on/off texts for the switch, and those texts
are translated. Russian language is known to be affected.

See a03e64af6db7023ccfcfaf29a825dc48b92c0c6a