Improve scrolling behavior of the MoveList area
Move the margin between the move list and the button row to the button
row container, so that the margin does not interfere with the
calculation in MoveListView.onMeasure().

Adjusted the MoveListView content size so that the last row can be
scrolled to a position where the previous line is completely hidden.