Make it possible to enter moves by dragging
Add new setting (default true) to allow input of moves by dragging a
piece from its source square to the target square.

This setting partially overlaps with scroll gestures on the chess
board, so when this setting is enabled, scroll gestures will only
happen if the starting square is empty or contains a piece of the
color not to make the next move.