Add file browser support when setting UCI options
For string UCI options containing "file" or "path" in their names, add
a button that opens a file browser to set the value of the option.

The browse button is only available if "OI file manager" or a
compatible app is installed. The button is not available for network
engines because there is no way to browse the remote filesystem where
the network engine runs.